requirements.txt
numpy
torch
Pillow
open_clip_torch
matplotlib
umap-learn
imageio
